requirements.txt
pymongo==4.5.0
pika==1.3.2
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.3.0
joblib==1.3.1
python-dotenv==1.0.0
requests
schedule==1.2.0
redis==4.5.5
